feat: Enterprise 필드 sales int -> Long 변경

+++ src/main/java/com/example/elancer/enterprise/dto/EnterpriseAccountDetailResponse.java
@@ -20,7 +20,7 @@ public class EnterpriseAccountDetailResponse {
     private String website;
     private Address address;
     private String bizContents;
-    private Integer sales;
+    private Long sales;
     private String idNumber;
 
     public static EnterpriseAccountDetailResponse of(Enterprise enterprise) {
